Revert "test download"
This reverts commit 72fce2f14bc21c9c14187b7472e2d02e4eb606f3.
